+++ crates/node-bindings/src/anvil.rs
@@ -119,6 +119,10 @@ pub enum AnvilError {
     /// An error occurred while parsing a hex string.
     #[error(transparent)]
     FromHexError(#[from] hex::FromHexError),
+
+    /// No keys available in anvil instance.
+    #[error("no keys available in anvil instance")]
+    NoKeysAvailable,
 }
 
 /// Builder for launching `anvil`.
@@ -314,7 +318,7 @@ impl Anvil {
 
         let mut child = cmd.spawn().map_err(AnvilError::SpawnError)?;
 
-        let stdout = child.stdout.as_mut().ok_or(AnvilError::NoStderr)?;
+        let stdout = child.stdout.take().ok_or(AnvilError::NoStderr)?;
 
         let start = Instant::now();
         let mut reader = BufReader::new(stdout);