+++ crates/rpc-types-eth/src/lib.rs
@@ -47,4 +47,7 @@ pub use transaction::*;
 mod work;
 pub use work::Work;
 
+/// This module provides implementations for EIP-4337.
+pub mod eip4337;
+
 pub mod simulate;